+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Engine.java
@@ -943,6 +943,8 @@ public final class Engine implements AutoCloseable {
          * @param value {@code true} to spawn a polyglot isolate
          * @see Context.Builder#spawnIsolate(boolean)
          * @see Engine#supportsCompilation()
+         * @see <a href="https://www.graalvm.org/latest/reference-manual/embed-languages/#polyglot-isolates">
+         *      Polyglot Isolates documentation</a>
          * @since 25.1
          */
         public Builder spawnIsolate(boolean value) {